Add a display module
Currently does little besides acting as a wrapper for print. In theory, compartmentalizing print makes it easier to add a graphical display, but you'd need to fix other issues as well.

Contains one function and a dummy display class

The function passes its input along to print or input, depending on whether or not it was told to pause after printing the output.
-Janky as all get-out
--end is ignored if pause is true
--the player can type (and see) input, which is ignored by the game.