Expand the example island to be visitable (but very very boring)
Locations are now Contexts (so they are expected to have verbs).
-Locations have a flag that controls if they are visitable or not (the pirates can go ashore).

SubLocations now exist (locations own sub-locations, which is important for connectivity).
-SubLocations are also contexts, so are expected to have verbs.

The island has two sub-locations: the beach and the forest. Nothing of note happens in either.

Changes how go is handled, so that it can be used for sailing the ship and exploring sub-locations.