Adjust combat code to include a defense option
A pirate that chooses to defend intercepts combat damage as it is applied to the defendee
-The defendee takes no damage AND the defender takes only 1/2 damage (rounded up).
--A pirate that chooses to defend another can still be attacked directly for 100% damage. Therefore, the expected damage is 150% of normal over time.
--A pirate that chooses to defend THEMSELVES takes only 50% damage.
-Prints a message as the interception occurs.

The inflict_damage function has been adjusted to return the true defender if they're dead, so that messages can be printed about them.

Pirates stop defending their target when their turn comes up in combat.